saga

+++ backend/domain/user-domain/src/main/java/org/atlas/domain/user/entity/ProductEntity.java
@@ -16,7 +16,7 @@ public class ProductEntity {
   private BigDecimal price;
   private String image;
 
-  public ProductEntity(Long id) {
+  public ProductEntity(Integer id) {
     this.id = id;
   }
 }
